Halved Drivetrain speed, made medium shoot mode
These are pretty bad fixes, but should work for now.  I halved the x and
y speeds in the drivetrain (NOT rotation), and set the teleop amp and
speaker commands to shoot at a medium speed. I also commented out the
shooting commands that rely on photonvision (since we don't have a
field) to avoid accidents